FIN-15 Importing colors globally. Putting constants into separate files.
