+++ src/src_V2/app/App.jsx
@@ -3,6 +3,8 @@ import { CssBaseline } from '@mui/material';
 
 import routesConfig from './routes';
 
+import '../styles/common.css';
+
 /**
  * @description I don't know why we need separate wrapper for useRoutes bu it didn't work as mentioned in doc.
  *              It's not final we need to check it out after.
