+++ src/src_V2/components/layout/Layout.module.css
@@ -1,5 +1,3 @@
-@import '../../styles/colors.css';
-
 .mainDiv {
   padding: 24px;
   height: 100%;
