+++ src/src_V2/components/loading-spinner/LoadingSpinner.module.css
@@ -1,5 +1,3 @@
-@import '../../styles/colors.css';
-
 .preloader {
   position: fixed;
   top: 0;
