+++ src/src_V2/components/no-data/NoData.module.css
@@ -1,5 +1,3 @@
-@import '../../styles/colors.css';
-
 .mainDiv {
   display: flex;
   flex-direction: column;
