+++ src/src_V2/components/sidebar/Sidebar.module.css
@@ -1,5 +1,3 @@
-@import '../../styles/colors.css';
-
 .mainBox {
   position: relative;
   display: flex;
