+++ src/src_V2/modules/manager/financial/components/payables/control-panel/ControlPanel.module.css
@@ -1,5 +1,3 @@
-@import '../../../../../../styles/colors.css';
-
 .mainDiv {
   display: flex;
   align-items: baseline;
