+++ src/src_V2/modules/manager/financial/components/payables/edit/drawer-content/RowList.module.css
@@ -1,5 +1,3 @@
-@import '../../../../../../../styles/colors.css';
-
 .mainText {
   font-size: 17px;
 }
