+++ src/src_V2/modules/manager/financial/components/payables/edit/drawer-content/payment-history/PaymentHistory.jsx
@@ -7,7 +7,7 @@ import {
   Typography,
 } from '@mui/material';
 
-import creditCard from '../../../../../../../../assets/credit-card.png';
+import creditCard from '../../../../../../../../assets/icons/credit-card.png';
 import commonStyles from '../style.module.css';
 import styles from './PaymentHistory.module.css';
 import { NoData } from '../../../../../../../../components';
