+++ src/src_V2/modules/manager/financial/components/payables/edit/drawer-content/style.module.css
@@ -1,5 +1,3 @@
-@import '../../../../../../../styles/colors.css';
-
 /* Accordions */
 
 .accordion {
