+++ src/src_V2/modules/manager/financial/components/payables/table/PayableTable.module.css
@@ -1,5 +1,3 @@
-@import '../../../../../../styles/colors.css';
-
 .tableContainer {
   overflow-y: auto;
   background-color: var(--midGrey) !important;
