--- src/src_V2/modules/manager/financial/constants/payment-type.js
@@ -1,8 +1,0 @@
-const PaymentType = {
-  Cash: 1,
-  Transfer: 2,
-  1: 'Cash',
-  2: 'Transfer',
-};
-
-export default PaymentType;
