+++ src/src_V2/modules/manager/financial/mock/payable.js
@@ -5,6 +5,7 @@ function createPayable(
   tourId,
   tourName,
   supplier,
+  currency,
   plannedCost,
   actualCost,
   paidCost,
@@ -25,6 +26,7 @@ function createPayable(
     tourId,
     tourName,
     supplier,
+    currency,
     plannedCost,
     actualCost,
     paidCost,
@@ -48,6 +50,7 @@ const payablesList = () => [
     18759,
     '5 day tour to Tsakhadzor',
     'Tatev Hotel',
+    'AMD',
     400,
     400,
     20,
@@ -71,6 +74,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     400,
     20,
@@ -90,6 +94,7 @@ const payablesList = () => [
     18759,
     'Very very looooooooooooooooooooooooooooooooooooong Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     400,
     20,
@@ -110,6 +115,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     300,
     20,
@@ -128,6 +134,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     500,
     20,
@@ -146,6 +153,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     400,
     20,
@@ -164,6 +172,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     600,
     1000,
     20,
@@ -182,6 +191,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     400,
     20,
@@ -200,6 +210,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     400,
     20,
@@ -218,6 +229,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     400,
     20,
@@ -236,6 +248,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     400,
     20,
@@ -254,6 +267,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     400,
     10,
     20,
@@ -272,6 +286,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     9000000000000,
     400,
     20,
@@ -290,6 +305,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     800,
     400,
     20,
@@ -308,6 +324,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     800,
     300,
     20,
@@ -326,6 +343,7 @@ const payablesList = () => [
     18759,
     'Trip to Tatev',
     'Tatev Hotel',
+    'AMD',
     1000,
     400,
     20,
