+++ src/src_V2/modules/manager/supplier-management/accommodation/add/style.module.css
@@ -1,5 +1,3 @@
-@import '../../../../../styles/colors.css';
-
 /* Add accommodation styles*/
 
 .container {
