+++ src/src_V2/styles/common.css
@@ -1,0 +1,3 @@
+@import 'colors.css';
+
+/* globally applied styles go here */
